Rename package to mongodb.
