+++ brave-bom/pom.xml
@@ -204,7 +204,7 @@
       </dependency>
       <dependency>
         <groupId>${project.groupId}</groupId>
-        <artifactId>brave-instrumentation-mongo</artifactId>
+        <artifactId>brave-instrumentation-mongodb</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
