+++ instrumentation/mongodb/README.md
@@ -1,6 +1,6 @@
-# brave-instrumentation-mongo
+# brave-instrumentation-mongodb
 
-This includes [`TraceMongoCommandListener`](src/main/java/brave/mongo/TraceMongoCommandListener.java), a
+This includes [`TraceMongoCommandListener`](src/main/java/brave/mongodb/TraceMongoCommandListener.java), a
 [`CommandListener`](https://mongodb.github.io/mongo-java-driver/3.12/driver/reference/monitoring/#command-monitoring)
 for the Mongo Java driver that will report via Brave how long each command takes, along with relevant tags like the
 database name, connection name, the command's operation (`insert`, `update`, `find`, etc.).
