+++ instrumentation/mongodb/bnd.bnd
@@ -1,4 +1,4 @@
 Import-Package: \
   *
 Export-Package: \
-  brave.mongo
+  brave.mongodb
