+++ instrumentation/mongodb/src/main/java/brave/mongodb/TraceMongoCommandListener.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package brave.mongo;
+package brave.mongodb;
 
 import brave.Span;
 import brave.Tracer;
